fix: componentizing property registry inputs
