+++ src/config/config.tsx
@@ -1,0 +1,14 @@
+const API = "https://any-where-back.vercel.app/";
+
+export const propertyRegister = async () => {
+    const bodyForm = JSON.stringify();
+    const config = {
+        method: "POST",
+        headers: {
+          "content-type": "application/json",
+        },
+        body: bodyForm,
+      };
+  return await fetch(`${API}/property`, config);
+
+}
